// src/generated/prisma/enums.ts
export const CareerLevel = {
  DIGITAL_ASSOCIATE: 'DIGITAL_ASSOCIATE',
  CERTIFIED_REVIEWER: 'CERTIFIED_REVIEWER',
  SENIOR_VERIFIER: 'SENIOR_VERIFIER',
  VERIFICATION_SPECIALIST: 'VERIFICATION_SPECIALIST',
  TEAM_SUPERVISOR: 'TEAM_SUPERVISOR'
} as const

export type CareerLevel = (typeof CareerLevel)[keyof typeof CareerLevel]
// ...
export const LeaderboardPeriod = {
  WEEKLY: 'WEEKLY',
  MONTHLY: 'MONTHLY',
  ALL_TIME: 'ALL_TIME'
} as const

export type LeaderboardPeriod = (typeof LeaderboardPeriod)[keyof typeof LeaderboardPeriod]